pyproject.toml
[tool.poetry]
name = "discotp"
version = "1.0.2"
description = "A Discord bot for generating and managing TOTPs (for shared accounts)."
authors = ["blobbybilb <notablob@outlook.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py-interactions = "^5.13.2"
sqlitedict = "^2.1.0"
pyotp = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
discotp = "discotp.run:main"